refactor(lint): no-unsafe-return + no-unsafe-argument -> error (W2 1b)
Promotes no-unsafe-return / no-unsafe-argument from warn to error in both the
engine-src and packages rule blocks, and removes them from the per-directory
off blocks (rendering/webgl2, rendering, math+core+debug+input, Loader,
IndexedDb+factories). The other no-unsafe-* (assignment/member-access/call)
stay warn/off -- out of scope.

Fixed the 14 findings at the root (no `as any` / disable):
- Array.isArray narrowing quirk (any[] erasure): annotate the local
  (Gamepad/InputManager InputChannel lists, AnimatedSprite frames, Loader paths).
- instanceof DataTexture erasing the format generic: annotate
  `format: DataTextureFormat` (WebGl2Backend, WebGpuBackend x2).
- untyped browser APIs returning any: WebGl2ShaderBlock getActiveUniform(Block)
  (Uint32Array/Int32Array per WebGL2 spec), IndexedDb IDBRequest result (typed
  record), WebGl2Backend Reflect.get (unknown + guard).

CI lint:all gates `error` (not `warn`), so this is now enforced.
lint:strict 0 - typecheck 0 - 492 + 74 affected tests green.

Spec: .workspace/specs/v0.15-hardening/03-typsystem.md (1b)

+++ src/rendering/webgl2/WebGl2ShaderBlock.ts
@@ -59,8 +59,12 @@ export class WebGl2ShaderBlock {
     const gl = this._context;
     const program = this._program;
     const blockData = this._blockData;
-    const indices = gl.getActiveUniformBlockParameter(program, this.index, gl.UNIFORM_BLOCK_ACTIVE_UNIFORM_INDICES);
-    const offsets = gl.getActiveUniforms(program, indices, gl.UNIFORM_OFFSET);
+    // `getActiveUniformBlockParameter`/`getActiveUniforms` are typed `any` in
+    // lib.dom; the WebGL2 spec returns a Uint32Array of active uniform indices
+    // for UNIFORM_BLOCK_ACTIVE_UNIFORM_INDICES and an Int32Array of byte offsets
+    // for UNIFORM_OFFSET.
+    const indices = gl.getActiveUniformBlockParameter(program, this.index, gl.UNIFORM_BLOCK_ACTIVE_UNIFORM_INDICES) as Uint32Array;
+    const offsets = gl.getActiveUniforms(program, indices, gl.UNIFORM_OFFSET) as Int32Array;
     const len = indices.length;
 
     for (let i = 0; i < len; i++) {
